Fixes memory leak
Previously, when one of the apss functions was called and the preprocess
function had not yet been called, the apss function called the
preprocess function and set a variable to make sure that the
preprocessed memory was released after computation. This worked, but the
preprocess struct itself was not being free'd in this case. This commit
fixes that.