Fixes jiverson002/efika-apss#1
The validate step in the allpairs implementation was depending on the
use of BLAS_vdoti() from the core component. That implementation
computes the vdot in forward order. Other implementations in this library
compute the final vdot in reverse order so as to apply filters and
reduce the number of mul+acc operations. This discrepancy led to
slightly different distances, due to floating-point precision, which in
turn led to slightly different counts of nearest-neighbors. This commit
replaces BLAS_vdoti() with the same reverse vdot that the other
implementations are using.